backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-limiter==3.8.0
itsdangerous==2.2.0
werkzeug==3.0.3
marshmallow==3.21.1
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
